Issue 480 omopgenerics type warinings (#490)
* sqlite is not supported

* suppress omopgenerics typecheck warnings. comment out achilles test until we get a test database with achilles tables.

* remove type check warning suppression. Switch back to using int instead of bigint for subject_id in cohort tables.

* remove type check warning supression in tests

* fix omopgenerics type check warnings

* fix type check warnings

* fix typo

* fix git action for code coverage

* debug code coverage github action

* add github workflows back

+++ R/cohort_ddl.R
@@ -39,8 +39,7 @@ createCohortTables <- function(con, writeSchema, name, computeAttrition) {
                      name = .inSchema(writeSchema, name, dbms(con)),
                      fields = c(
                        cohort_definition_id = "INT",
-                       subject_id = "BIGINT",
-                       # subject_id = "INT",
+                       subject_id = "INT",
                        cohort_start_date = "DATE",
                        cohort_end_date = "DATE"))
